Fix errors introduced in the webpack build
Aurelia dialog prevented the whole page from loading
because the dependencies of the plugin could not be correctly resolved
in the output from the webpack build.
The fix enforces chuck optimization for all chuncks
which dedupes code across all existing chunks.

Setting runtimeChunk: 'single' does not contribute to the fix but seemed
like a good idea anyway and hopefully prevents future problems.